refactor(vMix): extract to multiple classes
not perfect in terms of separation of concerns, but it works
